test(auto-reply): move fast-path sidebar-group case into its own file
get-reply.fast-path.test.ts was already at the 1000-line test budget, so the added case tripped eslint(max-lines) in check-lint. Keep the coverage in a focused sibling file.
